Changed int to double for interval

+++ dart_ping/lib/src/ping/base_ping.dart
@@ -25,7 +25,7 @@ abstract class BasePing {
   int? count;
 
   /// Delay between ping attempts
-  int interval;
+  double interval;
 
   /// How long to wait for a ping to return before marking it as lost
   int timeout;
